+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.3
+
+* Updated dependencies
+* Add support for iPhone 17 series: iPhone 17, iPhone Air, iPhone 17 Pro, iPhone 17 Pro Max
+* Optimize layout logic
+* Update README typos
+* Add more documentation
+* Add more tests
+
 ## 0.1.2
 
 * Updated README